+++ src/dray/dispatcher.hpp
@@ -6,8 +6,8 @@
 #ifndef DRAY_DISPATCHER_HPP
 #define DRAY_DISPATCHER_HPP
 
-#include<dray/topology_base.hpp>
-#include<dray/derived_topology.hpp>
+#include<dray/data_model/topology_base.hpp>
+#include<dray/data_model/derived_topology.hpp>
 #include<dray/data_model/field.hpp>
 #include<dray/error.hpp>
 #include<dray/utils/data_logger.hpp>
