+++ src/dray/dray_node_to_dataset.cpp
@@ -8,7 +8,7 @@
 #include <dray/data_model/elem_attr.hpp>
 #include <dray/data_model/grid_function.hpp>
 #include <dray/data_model/mesh.hpp>
-#include <dray/derived_topology.hpp>
+#include <dray/data_model/derived_topology.hpp>
 
 #include <string>
 #include <vector>
