+++ src/dray/filters/to_bernstein.cpp
@@ -12,8 +12,8 @@
 #include <dray/policies.hpp>
 #include <dray/exports.hpp>
 
-#include <dray/topology_base.hpp>
-#include <dray/derived_topology.hpp>
+#include <dray/data_model/topology_base.hpp>
+#include <dray/data_model/derived_topology.hpp>
 #include <dray/data_model/mesh.hpp>
 #include <dray/data_model/field.hpp>
 #include <dray/data_model/device_mesh.hpp>
